Comments and cleanup

+++ objects/player/type.go
@@ -4,7 +4,7 @@ type Player struct {
 	PlayerID             int
 	PositionX, PositionY float64
 	Velocity             int
-	Rotation             float64
+	Rotation             int
 	Health               int
 	Points               int
 	Deaths               int
